--- minijava-typechecker/picojava/TestTypeCheckSimp.java
@@ -1,10 +1,0 @@
-package picojava;
-
-import java.util.*;
-import syntaxtree.*;
-import visitor.*;
-
-public class TestTypeCheckSimp {
-    public TestTypeCheckSimp(){
-    }
-}
