+++ minijava-typechecker/picojava/TestTypeValidator.java
@@ -1,9 +1,10 @@
 package picojava;
+
 import java.util.*;
 import syntaxtree.*;
 import visitor.*;
 
-public class TestSymTableVis {
-    public TestSymTableVis(){
+public class TestTypeValidator {
+    public TestTypeValidator(){
     }
 }
