--- minijava-typechecker/picojava/TypeCheckSimp.java
@@ -1,7 +1,0 @@
-package picojava;
-import syntaxtree.*;
-import visitor.*;
-import java.util.HashMap;
-
-public class TypeCheckSimp<R, A> extends GJDepthFirst<MyType, HashMap<String, String>> {
-}
