+++ minijava-typechecker/picojava/TypeValidator.java
@@ -1,0 +1,7 @@
+package picojava;
+import syntaxtree.*;
+import visitor.*;
+import java.util.HashMap;
+
+public class TypeValidator<R, A> extends GJDepthFirst<MyType, HashMap<String, String>> {
+}
